add latest changes - pages are initialized from BasePage
