+++ src/main/java/pageObjects/BasePage.java
@@ -5,6 +5,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.openqa.selenium.*;
 import org.openqa.selenium.NoSuchElementException;
+import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 import org.openqa.selenium.support.ui.*;
@@ -132,7 +133,16 @@ public class BasePage extends Utils {
      * @param driver
      */
     public BasePage(WebDriver driver) {
-
+        if (driver == null) {
+            try {
+                GetBrowser();
+            } catch (Exception e) {
+                e.printStackTrace();
+            }
+            if (currentBrowser.toLowerCase().contains("chrome")) {
+                driver = new ChromeDriver();
+            }
+        }
         this.driver = driver;
         env = System.getProperty("run.env");
